Move extract method code to preload
