+++ src/global.d.ts
@@ -1,14 +1,14 @@
 import { PlatformPath } from "path";
 import os from "node:os";
 import fs from "fs";
-import { nodeAPI } from "entry/preload";
+import { mainAPI, nodeAPI } from "entry/preload";
 
 export {};
 
 declare global {
   interface Window {
     // Created in electron preload script
-    api: any;
+    api: mainAPI;
 
     node: nodeAPI;
